segmentation/提交测试/requirements.txt
albumentations==1.1.0
matplotlib==3.4.2
numpy==1.21.5
opencv_python_headless==4.5.5.64
Pillow~=9.0.1
segmentation_models_pytorch==0.2.1
torch==1.9.0
torchvision==0.11.3
tqdm==4.64.0

opencv-python~=4.5.5.64